server/requirements.txt
django==4.2.2
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.2
celery==5.2.7
django-celery==3.1.17
redis==4.0.2
google-api-python-client==2.90.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
